pyproject.toml
[tool.poetry]
name = "coopi"
version = "0.1.0"
description = ""
authors = ["Liam MacKenzie <lmacka@gmail.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "piwheels"
url = "https://www.piwheels.org/simple/"

[tool.poetry.dependencies]
python = ">=3.9"
imutils = "^0.5.4"
rpi-gpio = "^0.7.1"
automationhat = "^0.4.1"
logging = "^0.4.9.6"
board = "^1.0"
numpy = "^1.26.4"

#[build-system]
#requires = ["poetry-core"]
#build-backend = "poetry.core.masonry.api"
smbus = "^1.1.post2"
rpi-ws281x = "4.3.4"
adafruit-circuitpython-bh1750 = "^1.1.10"
